migrate amounts from int to BigInteger

+++ src/main/java/org/dsaw/poker/engine/Client.java
@@ -17,9 +17,9 @@
 
 package org.dsaw.poker.engine;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Set;
-
 import org.dsaw.poker.engine.actions.Action;
 
 /**
@@ -51,7 +51,7 @@ public interface Client {
      * @param players
      *            The players at the table (including this player).
      */
-    void joinedTable(TableType type, int bigBlind, List<Player> players);
+    void joinedTable(TableType type, BigInteger bigBlind, List<Player> players);
     
     /**
      * Handles the start of a new hand.
@@ -87,7 +87,7 @@ public interface Client {
      * @param pot
      *            The current pot.
      */
-    void boardUpdated(List<Card> cards, int bet, int pot);
+    void boardUpdated(List<Card> cards, BigInteger bet, BigInteger pot);
     
     /**
      * Handles the event of a player acting.
@@ -109,6 +109,6 @@ public interface Client {
      * 
      * @return The selected action.
      */
-    Action act(int minBet, int currentBet, Set<Action> allowedActions);
+    Action act(BigInteger minBet, BigInteger currentBet, Set<Action> allowedActions);
 
 }
